fix: Resolved some  comments , Updater Function Remaining to move in ViewModel [123]
