+++ src/packages/@common/components/index.ts
@@ -1,3 +1,3 @@
 export { default as Sidebar } from "./sidebar/Sidebar.svelte";
 export { default as Header } from "./header/Header.svelte";
-export { default as Ticker } from "./ticker/Ticker.svelte";
+export { default as Updater } from "./updater/Updater.svelte";
