--- src/packages/@library/icons/CrossIconWhite.svelte
@@ -1,3 +1,0 @@
-<svg width="12" height="16" viewBox="0 0 12 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M10.7062 4.70469C11.0968 4.31406 11.0968 3.67969 10.7062 3.28906C10.3155 2.89844 9.68115 2.89844 9.29053 3.28906L5.9999 6.58281L2.70615 3.29219C2.31553 2.90156 1.68115 2.90156 1.29053 3.29219C0.899902 3.68281 0.899902 4.31719 1.29053 4.70781L4.58428 7.99844L1.29365 11.2922C0.903027 11.6828 0.903027 12.3172 1.29365 12.7078C1.68428 13.0984 2.31865 13.0984 2.70928 12.7078L5.9999 9.41406L9.29365 12.7047C9.68428 13.0953 10.3187 13.0953 10.7093 12.7047C11.0999 12.3141 11.0999 11.6797 10.7093 11.2891L7.41553 7.99844L10.7062 4.70469Z" fill="#F5F5F5"/>
-</svg>
