+++ src/packages/@support/common/constants/learnMore.constant.ts
@@ -1,1 +1,2 @@
-export const LearnMoreURL = "https://github.com/sparrowapp-dev/sparrow-app/releases";
+export const LearnMoreURL =
+  "https://github.com/sparrowapp-dev/sparrow-app/releases";
